Add Limitation for change status to visited or canceled

+++ app/Http/Controllers/FinancialTransactionController.php
@@ -76,7 +76,6 @@ class FinancialTransactionController extends Controller
 
     public function store(FinancialTransactionRequest $request)
     {
-        FinancialTransaction::create($request->all());
         if ($request->input("appointment_id") !== null){
             $appointment_id=$request->input("appointment_id");
             $appointment=Appointment::findOrFail($appointment_id);
@@ -85,12 +84,14 @@ class FinancialTransactionController extends Controller
             $patientN=$appointment->patient;
             $patient_id=$appointment->patient->national_code;
             $appointment=new Appointment();
+            FinancialTransaction::create([...$request->all(),'changeable'=>false]);
             return view('admin.prescriptions.add-level2',[
                 'appointments'=>$appointment->where("patient_id",$patient_id)->whereIn("status",[0,1])->orderby("visit_time")->get(),
                 'patient'=>$patientN,
                 'appointment_id'=>$appointment_id,
             ]);
         }else{
+            FinancialTransaction::create($request->all());
             return redirect()->route('financials');
         }
     }
@@ -102,8 +103,9 @@ class FinancialTransactionController extends Controller
         ]);
     }
 
-    public function edit(FinancialTransaction $financialTransaction): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application
+    public function edit(FinancialTransaction $financialTransaction): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application| \Illuminate\Http\RedirectResponse
     {
+        if ($financialTransaction->changeable){
         $patient= new Patient;
         $methods = [ "دستگاه کارتخوان" , "کارت به کارت", "نقدی" , "چندحالتی" , "غیره"];
         return view('admin.financial_transactions.edit',[
@@ -111,17 +113,25 @@ class FinancialTransactionController extends Controller
             'patients'=>$patient->orderBy("firstname","asc")->orderBy("lastname","asc")->get(),
             'financialTransaction'=>$financialTransaction,
         ]);
+        }else{
+            return redirect()->route('financials');
+        }
+
     }
 
     public function update(FinancialTransactionRequest $request,FinancialTransaction $financialTransaction): \Illuminate\Http\RedirectResponse
     {
-        $financialTransaction->update($request->all());
+        if ($financialTransaction->changeable){
+            $financialTransaction->update($request->all());
+        }
         return redirect()->route('financials');
     }
 
     public function destroy(FinancialTransaction $financialTransaction): \Illuminate\Http\RedirectResponse
     {
-        $financialTransaction->delete();
+        if ($financialTransaction->changeable){
+            $financialTransaction->delete();
+        }
         return redirect()->route('financials');
     }
